+++ src/index.js
@@ -166,7 +166,7 @@ if (typeof window !== 'undefined' && window.Vue) {
 }
 
 module.exports = {
-  version: '1.1.14',
+  version: '1.1.15',
   locale: locale.use,
   i18n: locale.i18n,
   install,